Modified buildMoveDatagram to take a const char* to allow for bounds checking before sending move and added bounds checking.
Made buildChatDatagram and inforced that the resulting datagram is 80 chars or less and starts with a 'C'.

Made buildForfeitDatagram.

Made promptAndSendNextDecision which takes a const char*. It prompts the user to make a move, send a chat or forfeit. It also allows the user to send as many messages as they want as long as they are less than 80 chars. It makse sure that the use picks a valid option.